Relocate the Modular Exponentiation problem under the Maths section
